added DuPPaLib for I2CNavKey

+++ arduino/WyoBean/textures.cpp
@@ -1,6 +1,7 @@
 #include "textures.h"
 
-void rainbow(CRGB *leds, uint16_t current_time, uint16_t (*XY)(uint8_t col, uint8_t row)) {
+void rainbow(CRGB *leds, uint32_t current_time, uint16_t (*XY)(int col, int row)) {
+  //void rainbow(CRGB *leds, uint16_t current_time, uint16_t (*XY)(uint8_t col, uint8_t row)) {
   int i, dx, dy;
   CHSV hsv;
   float dist;
