+++ arduino/WyoBean/textures.h
@@ -3,7 +3,8 @@
 
 #include <FastLED.h>
 
-void rainbow(CRGB *leds, uint16_t current_time, uint16_t (*XY)(uint8_t col, uint8_t row));
+void rainbow(CRGB *leds, uint32_t current_time, uint16_t (*XY)(int col, int row));
+//void rainbow(CRGB *leds, uint16_t current_time, uint16_t (*XY)(uint8_t col, uint8_t row));
 void noop();
 void blend_to_rainbow();
 void rainbow();
